+++ src/analyze-action.ts
@@ -15,7 +15,6 @@ import {
 import { CODEQL_VERSION_NEW_TRACING, getCodeQL } from "./codeql";
 import { Config, getConfig } from "./config-utils";
 import { uploadDatabases } from "./database-upload";
-import { GitHubFeatureFlags } from "./feature-flags";
 import { getActionsLogger } from "./logging";
 import { parseRepositoryNwo } from "./repository";
 import * as upload_lib from "./upload-lib";
@@ -36,6 +35,7 @@ interface FinishStatusReport
 
 export async function sendStatusReport(
   startedAt: Date,
+  config: Config | undefined,
   stats: AnalysisStatusReport | undefined,
   error?: Error
 ) {
@@ -52,6 +52,12 @@ export async function sendStatusReport(
   );
   const statusReport: FinishStatusReport = {
     ...statusReportBase,
+    ...(config
+      ? {
+          ml_powered_javascript_queries:
+            util.getMlPoweredJsQueriesStatus(config),
+        }
+      : {}),
     ...(stats || {}),
   };
   await actionsUtil.sendStatusReport(statusReport);
@@ -105,12 +111,6 @@ async function run() {
       util.getRequiredEnvParam("GITHUB_REPOSITORY")
     );
 
-    const featureFlags = new GitHubFeatureFlags(
-      config.gitHubVersion,
-      apiDetails,
-      repositoryNwo,
-      logger
-    );
     await runFinalize(outputDir, threads, memory, config, logger);
     if (actionsUtil.getRequiredInput("skip-queries") !== "true") {
       runStats = await runQueries(
@@ -193,13 +193,7 @@ async function run() {
     }
 
     // Possibly upload the database bundles for remote queries
-    await uploadDatabases(
-      repositoryNwo,
-      config,
-      featureFlags,
-      apiDetails,
-      logger
-    );
+    await uploadDatabases(repositoryNwo, config, apiDetails, logger);
 
     if (
       uploadResult !== undefined &&
@@ -220,9 +214,9 @@ async function run() {
 
     if (error instanceof CodeQLAnalysisError) {
       const stats = { ...error.queriesStatusReport };
-      await sendStatusReport(startedAt, stats, error);
+      await sendStatusReport(startedAt, config, stats, error);
     } else {
-      await sendStatusReport(startedAt, undefined, error);
+      await sendStatusReport(startedAt, config, undefined, error);
     }
 
     return;
@@ -279,14 +273,14 @@ async function run() {
   }
 
   if (runStats && uploadResult) {
-    await sendStatusReport(startedAt, {
+    await sendStatusReport(startedAt, config, {
       ...runStats,
       ...uploadResult.statusReport,
     });
   } else if (runStats) {
-    await sendStatusReport(startedAt, { ...runStats });
+    await sendStatusReport(startedAt, config, { ...runStats });
   } else {
-    await sendStatusReport(startedAt, undefined);
+    await sendStatusReport(startedAt, config, undefined);
   }
 }