+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CodeQL Action and CodeQL Runner Changelog
 
+## 1.1.1 - 17 Feb 2022
+
+- The CodeQL CLI versions up to and including version 2.4.4 are not compatible with the CodeQL Action 1.1.1 and later. The Action will emit an error if it detects that it is being used by an incompatible version of the CLI. [#931](https://github.com/github/codeql-action/pull/931)
+- Update default CodeQL bundle version to 2.8.1. [#925](https://github.com/github/codeql-action/pull/925)
+
 ## 1.1.0 - 11 Feb 2022
 
 - The CodeQL Action now uses Node.js v16. [#909](https://github.com/github/codeql-action/pull/909)